Intermediate commit Changed offset to be isize Added some tests Need to add ability to process mixing symbols such as '[foo]?' or '[foo]*', which is not possible at the moment.

+++ src/regex/qst.rs
@@ -40,7 +40,7 @@ impl RegexQst {
 }
 
 impl RegexToken for RegexQst {
-    fn extract_text(&mut self, txt: &str, offset: i32) -> Option<TextExtract> {
+    fn extract_text(&mut self, txt: &str, offset: isize) -> Option<TextExtract> {
         None
     }
